creating atomic components
